pam: Move C code to separated files, to keep it cleaner and easy to maintain
